+++ test/services/husqvarna/discoveryService.spec.ts
@@ -1,12 +1,12 @@
 import { It, Mock, Times } from 'moq.ts';
 
-import { AutomowerAccessory } from '../../../../src/automowerAccessory';
-import { AutomowerAccessoryFactory } from '../../../../src/automowerAccessoryFactory';
-import { AutomowerPlatform } from '../../../../src/automowerPlatform';
-import { Activity, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../../src/clients/automowerClient';
-import { PlatformLogger } from '../../../../src/diagnostics/platformLogger';
-import { DiscoveryServiceImpl } from '../../../../src/services/husqvarna/automower/discoveryService';
-import { GetMowersService } from '../../../../src/services/husqvarna/automower/getMowersService';
+import { AutomowerAccessory } from '../../../src/automowerAccessory';
+import { AutomowerAccessoryFactory } from '../../../src/automowerAccessoryFactory';
+import { AutomowerPlatform } from '../../../src/automowerPlatform';
+import { Activity, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../src/clients/automowerClient';
+import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
+import { GetMowersService } from '../../../src/services/husqvarna/automower/getMowersService';
+import { DiscoveryServiceImpl } from '../../../src/services/husqvarna/discoveryService';
 
 describe('DiscoveryServiceImpl', () => {
     let getMowersService: Mock<GetMowersService>;
